move topic listing route from NothingPersonalHomePage to HomePage
