+++ foundation_cms/templates/patterns/pages/core/topic_listing_page.html
@@ -9,6 +9,8 @@
   <link rel="stylesheet" href="{% static 'foundation_cms/_css/pages/topic_listing_page.compiled.css' %}">
 {% endblock extra_css %}
 
+{% block body_class %}topic-listing-page{% endblock body_class %}
+
 {% block pre_content %}{% endblock pre_content %}
 
 {% block content %}
